+++ challenge/Posts.swift
@@ -12,9 +12,9 @@ import CoreData
 class Posts: NSManagedObject {
     
     @NSManaged public var body: String
-    @NSManaged public var id: Int64
+    @NSManaged public var id: Int
     @NSManaged public var title: String
-    @NSManaged public var userId: Int64
+    @NSManaged public var userId: Int
     @NSManaged public var user: User
     
 }
