+++ challenge/User.swift
@@ -6,7 +6,7 @@ class User: NSManagedObject {
     @NSManaged public var address: NSData
     @NSManaged public var company: NSData
     @NSManaged public var email: String
-    @NSManaged public var id: Int64
+    @NSManaged public var id: Int
     @NSManaged public var name: String
     @NSManaged public var phone: String
     @NSManaged public var username: String
